refactor(cargo): Pull out sort logic
