+++ src/cargo.rs
@@ -203,6 +203,56 @@ pub fn parse_cargo_config(manifest_path: &Path) -> Result<Value, FatalError> {
     cargo_file_content.parse().map_err(FatalError::from)
 }
 
+pub fn sort_workspace(ws_meta: &cargo_metadata::Metadata) -> Vec<&cargo_metadata::PackageId> {
+    let members: std::collections::HashSet<_> = ws_meta.workspace_members.iter().collect();
+    let dep_tree: std::collections::HashMap<_, _> = ws_meta
+        .resolve
+        .as_ref()
+        .expect("cargo-metadata resolved deps")
+        .nodes
+        .iter()
+        .filter_map(|n| {
+            if members.contains(&n.id) {
+                Some((&n.id, &n.dependencies))
+            } else {
+                None
+            }
+        })
+        .collect();
+
+    let mut sorted = Vec::new();
+    let mut processed = std::collections::HashSet::new();
+    for pkg_id in ws_meta.workspace_members.iter() {
+        sort_workspace_inner(ws_meta, pkg_id, &dep_tree, &mut processed, &mut sorted);
+    }
+
+    sorted
+}
+
+fn sort_workspace_inner<'m>(
+    ws_meta: &'m cargo_metadata::Metadata,
+    pkg_id: &'m cargo_metadata::PackageId,
+    dep_tree: &std::collections::HashMap<
+        &'m cargo_metadata::PackageId,
+        &'m Vec<cargo_metadata::PackageId>,
+    >,
+    processed: &mut std::collections::HashSet<&'m cargo_metadata::PackageId>,
+    sorted: &mut Vec<&'m cargo_metadata::PackageId>,
+) {
+    if !processed.insert(pkg_id) {
+        return;
+    }
+
+    for dep_id in dep_tree[pkg_id]
+        .iter()
+        .filter(|dep_id| dep_tree.contains_key(dep_id))
+    {
+        sort_workspace_inner(ws_meta, dep_id, dep_tree, processed, sorted);
+    }
+
+    sorted.push(pkg_id);
+}
+
 #[cfg(test)]
 mod test {
     use super::*;
