+++ src/main.rs
@@ -1,5 +1,4 @@
 use std::collections::HashMap;
-use std::collections::HashSet;
 use std::ffi::OsStr;
 use std::io::Write;
 use std::path::Path;
@@ -373,7 +372,7 @@ fn release_workspace(args: &ReleaseOpt) -> Result<i32, error::FatalError> {
         release_config
     };
 
-    let pkg_ids = sort_workspace(&ws_meta);
+    let pkg_ids = cargo::sort_workspace(&ws_meta);
 
     let (selected_pkgs, excluded_pkgs) = args.workspace.partition_packages(&ws_meta);
     if selected_pkgs.is_empty() {
@@ -397,53 +396,6 @@ fn release_workspace(args: &ReleaseOpt) -> Result<i32, error::FatalError> {
         .collect();
 
     release_packages(args, &ws_meta, &ws_config, pkg_releases.as_slice())
-}
-
-fn sort_workspace(ws_meta: &cargo_metadata::Metadata) -> Vec<&cargo_metadata::PackageId> {
-    let members: HashSet<_> = ws_meta.workspace_members.iter().collect();
-    let dep_tree: HashMap<_, _> = ws_meta
-        .resolve
-        .as_ref()
-        .expect("cargo-metadata resolved deps")
-        .nodes
-        .iter()
-        .filter_map(|n| {
-            if members.contains(&n.id) {
-                Some((&n.id, &n.dependencies))
-            } else {
-                None
-            }
-        })
-        .collect();
-
-    let mut sorted = Vec::new();
-    let mut processed = HashSet::new();
-    for pkg_id in ws_meta.workspace_members.iter() {
-        sort_workspace_inner(ws_meta, pkg_id, &dep_tree, &mut processed, &mut sorted);
-    }
-
-    sorted
-}
-
-fn sort_workspace_inner<'m>(
-    ws_meta: &'m cargo_metadata::Metadata,
-    pkg_id: &'m cargo_metadata::PackageId,
-    dep_tree: &HashMap<&'m cargo_metadata::PackageId, &'m std::vec::Vec<cargo_metadata::PackageId>>,
-    processed: &mut HashSet<&'m cargo_metadata::PackageId>,
-    sorted: &mut Vec<&'m cargo_metadata::PackageId>,
-) {
-    if !processed.insert(pkg_id) {
-        return;
-    }
-
-    for dep_id in dep_tree[pkg_id]
-        .iter()
-        .filter(|dep_id| dep_tree.contains_key(dep_id))
-    {
-        sort_workspace_inner(ws_meta, dep_id, dep_tree, processed, sorted);
-    }
-
-    sorted.push(pkg_id);
 }
 
 fn release_packages<'m>(
